requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
dagviz==0.3.0
discord==2.1.0
discord.py==2.1.0
exceptiongroup==1.0.4
fonttools==4.38.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.3
networkx==2.8.8
numpy==1.23.5
opencv-contrib-python==4.6.0.66
outcome==1.2.0
packaging==23.2
Pillow==10.1.0
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
pytweening==1.0.7
requests==2.28.1
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
svgwrite==1.4.3
toml==0.10.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.9.0
urllib3==1.26.13
vboxapi==1.0
wsproto==1.2.0
yarl==1.8.2
